+++ application/models/Importexport.php
@@ -2807,18 +2807,9 @@ class Importexport extends ImportexportCommon
                     $rs = $srch->getResultSet();
                     $res = FatApp::getDb()->fetchAll($rs);
                     foreach ($res as $val) {
-                        $langSrch = new SearchBase(Product::DB_PRODUCT_LANG_SPECIFICATION);
-                        $langSrch->addCondition('prodspeclang_prodspec_id', '=', $val['prodspec_id']);
-                        $langSrch->addCondition('prodspeclang_lang_id', '!=', $langId);
-                        $langSrch->doNotCalculateRecords();
-                        $langSrch->addMultipleFields(['count(1) as record']);
-                        $langRs = $langSrch->getResultSet();
-                        $row = FatApp::getDb()->fetch($langRs);
-                        if (false == $row || $row['record'] == 0) {
-                            $this->db->deleteRecords(Product::DB_PRODUCT_SPECIFICATION, array('smt' => 'prodspec_product_id = ? ', 'vals' => array($productId)));
-                            $this->db->deleteRecords(Product::DB_PRODUCT_LANG_SPECIFICATION, array('smt' => 'prodspeclang_prodspec_id = ?', 'vals' => array($val['prodspec_id'])));
-                        } else {
-                            $this->db->deleteRecords(Product::DB_PRODUCT_LANG_SPECIFICATION, array('smt' => 'prodspeclang_prodspec_id = ? and prodspeclang_lang_id = ?', 'vals' => array($val['prodspec_id'], $langId)));
+                        $prodSpec = new ProdSpecification($val['prodspec_id']);
+                        if (!$prodSpec->deleteRecords($langId)) {
+                            LibHelper::exitWithError($prodSpec->getError(), true);
                         }
                     }
                 }